+++ compiler/src/iree/compiler/Codegen/VMVXPasses.h
@@ -1,0 +1,57 @@
+// Copyright 2023 The IREE Authors
+//
+// Licensed under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//===----------------------------------------------------------------------===//
+//
+// This file includes the VMVX related Passes.
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef IREE_COMPILER_CODEGEN_VMVX_PASSES_H_
+#define IREE_COMPILER_CODEGEN_VMVX_PASSES_H_
+
+#include "iree/compiler/Codegen/Dialect/LoweringConfig.h"
+#include "iree/compiler/Codegen/Utils/Utils.h"
+#include "iree/compiler/Dialect/HAL/IR/HALOps.h"
+#include "mlir/Dialect/Bufferization/IR/BufferizableOpInterface.h"
+#include "mlir/Dialect/Func/IR/FuncOps.h"
+#include "mlir/Dialect/Linalg/Transforms/Transforms.h"
+#include "mlir/Pass/Pass.h"
+#include "mlir/Pass/PassOptions.h"
+#include "mlir/Support/LLVM.h"
+#include "mlir/Transforms/DialectConversion.h"
+
+namespace mlir {
+namespace iree_compiler {
+//------------------------------------------------------------------------------
+// VMVX passes
+//------------------------------------------------------------------------------
+
+/// Materialize the encoding of operations. The layout to use for the encoded
+/// operations are VMVX specific.
+std::unique_ptr<OperationPass<func::FuncOp>>
+createVMVXMaterializeEncodingPass();
+
+// Lowers high level library calls from named ops and generics. This operates
+// at the bufferized linalg level.
+std::unique_ptr<Pass> createVMVXLowerLinalgMicrokernelsPass();
+
+//----------------------------------------------------------------------------//
+// VMVX Linking Passes and Pipelines
+//----------------------------------------------------------------------------//
+
+/// Links VMVX HAL executables within the top-level program module.
+std::unique_ptr<OperationPass<mlir::ModuleOp>> createVMVXLinkExecutablesPass();
+
+/// Assigns executable constant ordinals across all VMVX variants.
+std::unique_ptr<OperationPass<IREE::HAL::ExecutableVariantOp>>
+createVMVXAssignConstantOrdinalsPass();
+
+/// Populates passes needed to link HAL executables across VMVX targets.
+void buildVMVXLinkingPassPipeline(OpPassManager &passManager);
+}  // namespace iree_compiler
+}  // namespace mlir
+
+#endif  // IREE_COMPILER_CODEGEN_VMVX_PASSES_H_
